:recycle: UI 컴포넌트 분리
