+++ src/shared/ui/button/Button.tsx
@@ -1,0 +1,10 @@
+import { forwardRef } from "react"
+
+import { buttonVariants } from "./variants"
+import { ButtonProps } from "./types"
+
+export const Button = forwardRef<HTMLButtonElement, ButtonProps>(({ className, variant, size, ...props }, ref) => {
+  return <button className={buttonVariants({ variant, size, className })} ref={ref} {...props} />
+})
+
+Button.displayName = "Button"
