+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,10 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export { DialogContent } from "./DialogContent"
+export { DialogHeader } from "./DialogHeader"
+export { DialogTitle } from "./DialogTitle"
+
+export const Dialog = DialogPrimitive.Root
+export const DialogTrigger = DialogPrimitive.Trigger
+export const DialogPortal = DialogPrimitive.Portal
+export const DialogOverlay = DialogPrimitive.Overlay
