+++ src/shared/ui/select/index.ts
@@ -1,0 +1,9 @@
+import * as SelectPrimitive from "@radix-ui/react-select"
+
+export { SelectContent } from "./SelectContent"
+export { SelectItem } from "./SelectItem"
+export { SelectTrigger } from "./SelectTrigger"
+
+export const Select = SelectPrimitive.Root
+export const SelectGroup = SelectPrimitive.Group
+export const SelectValue = SelectPrimitive.Value
